Update library.js
Forcing the page to scroll to the bottom when the book form is opened to direct attention there for the user. Changing how I convert the items on line 65 and 82 so that it is more readable. Adding data-array-num to the book tiles/cards. Changing the books so that they feel more neutral.